etl/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
pandas>=2.1.0
pyarrow>=14.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
lxml>=4.9.0
duckdb>=0.10.0
homeharvest>=0.3.6
